¸üÐÂ

+++ exercises/functions/functions4.rs
@@ -23,6 +23,6 @@ fn sale_price(price: i32) -> bool{
     }
 }
 
-fn is_even(num: i32) -> u32 {
+fn is_even(num: i32) -> f32 {
     num % 2 == 0
 }
